Siv3D0.6.12から0.6.14にバージョンアップ

+++ Fumichou-Nes/src/Cartridge_In.cpp
@@ -31,7 +31,7 @@ private:
 			return CreateBoardNrom(rom);
 
 		default:
-			Logger::Error(fmt::format(U"Unsupported mapper: {}", mapper));
+			Logger::Error(fmt::format(U"Unsupported mapper: {}", static_cast<uint8>(mapper)));
 			return CreateBoardNrom(rom);
 		}
 	}
